requirements.txt
python-dotenv>=1.0.0
pyautogen>=0.2.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
unstructured>=0.10.0
markdown>=3.4.3
python-magic-bin>=0.4.14
pydantic>=2.0.0
langgraph>=0.0.15
tiktoken>=0.5.1
openai>=1.0.0
chromadb
pypdf
